Add web implementation of new app/window APIs.
