+++ web/src/toga_web/app.py
@@ -1,32 +1,37 @@
 import toga
-from toga.command import Separator
 from toga_web.libs import create_element, js
-from toga_web.window import Window
 
 from .screens import Screen as ScreenImpl
 
 
-class MainWindow(Window):
-    def on_close(self, *args):
-        pass
-
-
 class App:
+    # Web apps exit when the last window is closed
+    CLOSE_ON_LAST_WINDOW = True
+
     def __init__(self, interface):
         self.interface = interface
         self.interface._impl = self
 
     def create(self):
-        # self.resource_path = os.path.dirname(os.path.dirname(NSBundle.mainBundle.bundlePath))
         self.native = js.document.getElementById("app-placeholder")
+        # self.resource_path = os.path.dirname(os.path.dirname(NSBundle.mainBundle.bundlePath))
 
-        formal_name = self.interface.formal_name
+        # Call user code to populate the main window
+        self.interface._startup()
 
+    ######################################################################
+    # Commands and menus
+    ######################################################################
+
+    def _menu_about(self, command, **kwargs):
+        self.interface.about()
+
+    def create_app_commands(self):
         self.interface.commands.add(
             # ---- Help menu ----------------------------------
             toga.Command(
                 self._menu_about,
-                "About " + formal_name,
+                f"About {self.interface.formal_name}",
                 group=toga.Group.HELP,
             ),
             toga.Command(
@@ -36,111 +41,9 @@ class App:
             ),
         )
 
-        # Create the menus. This is done before main window content to ensure
-        # the <header> for the menubar is inserted before the <main> for the
-        # main window.
-        self.create_menus()
-
-        # Call user code to populate the main window
-        self.interface._startup()
-
-    ######################################################################
-    # Commands and menus
-    ######################################################################
-
-    def _create_submenu(self, group, items):
-        submenu = create_element(
-            "sl-dropdown",
-            children=[
-                create_element(
-                    "span",
-                    id=f"menu-{id(group)}",
-                    classes=["menu"],
-                    slot="trigger",
-                    content=group.text,
-                ),
-                create_element(
-                    "sl-menu",
-                    children=items,
-                ),
-            ],
-        )
-        return submenu
-
-    def _menu_about(self, command, **kwargs):
-        self.interface.about()
-
     def create_menus(self):
-        self._menu_groups = {}
-        submenu = None
-
-        for cmd in self.interface.commands:
-            if isinstance(cmd, Separator):
-                # TODO - add a section break
-                pass
-            else:
-                # TODO - this doesn't handle submenus properly;
-                # all menu items will appear in their root group.
-                submenu = self._menu_groups.setdefault(cmd.group, [])
-
-                menu_item = create_element(
-                    "sl-menu-item",
-                    content=cmd.text,
-                    disabled=not cmd.enabled,
-                )
-                menu_item.onclick = cmd._impl.dom_click
-
-                submenu.append(menu_item)
-
-        menu_container = create_element("nav", classes=["menubar"])
-        help_menu_container = create_element("nav", classes=["menubar"])
-
-        # If there isn't an explicit app menu group, add an inert placeholder
-        if toga.Group.APP not in self._menu_groups:
-            menu_container.appendChild(
-                create_element(
-                    "span",
-                    classes=["app"],
-                    content=self.interface.formal_name,
-                )
-            )
-
-        for group, items in self._menu_groups.items():
-            submenu = self._create_submenu(group, items)
-            if group != toga.Group.HELP:
-                menu_container.appendChild(submenu)
-            else:
-                help_menu_container.appendChild(submenu)
-
-        menubar_id = f"{self.interface.app_id}-header"
-        self.menubar = create_element(
-            "header",
-            id=menubar_id,
-            classes=["toga"],
-            children=[
-                create_element(
-                    "a",
-                    classes=["brand"],
-                    children=[
-                        create_element(
-                            "img",
-                            src="static/logo-32.png",
-                            alt=f"{self.interface.formal_name} logo",
-                            loading="lazy",
-                        )
-                    ],
-                ),
-                menu_container,
-                help_menu_container,
-            ],
-        )
-
-        # If there's an existing menubar, replace it.
-        old_menubar = js.document.getElementById(menubar_id)
-        if old_menubar:
-            old_menubar.replaceWith(self.menubar)
-        else:
-            self.native.append(self.menubar)
+        for window in self.interface.windows:
+            window._impl.create_menus()
 
     ######################################################################
     # App lifecycle
@@ -149,11 +52,18 @@ class App:
     def exit(self):
         pass
 
+    def finalize(self):
+        self.create_app_commands()
+        self.create_menus()
+
     def main_loop(self):
         self.create()
 
     def set_main_window(self, window):
-        pass
+        if window is None:
+            raise RuntimeError("Session-based apps are not supported on Web")
+        elif window == toga.App.BACKGROUND:
+            raise RuntimeError("Background apps are not supported on Web")
 
     ######################################################################
     # App resources
