+++ web/src/toga_web/factory.py
@@ -1,7 +1,7 @@
 from toga import NotImplementedWarning
 
 from . import dialogs
-from .app import App, MainWindow  # DocumentApp
+from .app import App
 from .command import Command
 
 # from .documents import Document
@@ -37,7 +37,7 @@ from .widgets.textinput import TextInput
 
 # from .widgets.tree import Tree
 # from .widgets.webview import WebView
-# from .window import Window
+from .window import MainWindow, Window
 
 
 def not_implemented(feature):
@@ -47,8 +47,6 @@ def not_implemented(feature):
 __all__ = [
     "not_implemented",
     "App",
-    "MainWindow",
-    # 'DocumentApp',
     "Command",
     # 'Document',
     # # Resources
@@ -80,7 +78,9 @@ __all__ = [
     "TextInput",
     # 'Tree',
     # 'WebView',
-    # 'Window',
+    # Windows
+    "MainWindow",
+    "Window",
 ]
