+++ web/src/toga_web/window.py
@@ -1,3 +1,5 @@
+import toga
+from toga.command import Separator
 from toga_web.libs import create_element, js
 
 from .screens import Screen as ScreenImpl
@@ -14,9 +16,6 @@ class Window:
             classes=["toga", "window", "container"],
             role="main",
         )
-
-        app_placeholder = js.document.getElementById("app-placeholder")
-        app_placeholder.appendChild(self.native)
 
         self.set_title(title)
 
@@ -47,11 +46,15 @@ class Window:
     def close(self):
         self.interface.factory.not_implemented("Window.close()")
 
-    def create_toolbar(self):
-        self.interface.factory.not_implemented("Window.create_toolbar()")
+    def create_menus(self):
+        # Simple windows don't have menus
+        pass
 
     def set_app(self, app):
-        pass
+        if len(app.interface.windows) > 1:
+            raise RuntimeError("Secondary windows cannot be created on Web")
+
+        app.native.appendChild(self.native)
 
     def show(self):
         self.native.style = "visibility: visible;"
@@ -121,3 +124,106 @@ class Window:
 
     def get_image_data(self):
         self.interface.factory.not_implemented("Window.get_image_data()")
+
+
+class MainWindow(Window):
+    def __init__(self, interface, title, position, size):
+        super().__init__(interface, title, position, size)
+
+        # Create the window titlebar, with placeholders for the menu items.
+        self.native_menu_container = create_element("nav", classes=["menubar"])
+        self.native_help_menu_container = create_element("nav", classes=["menubar"])
+        self.native_titlebar = create_element(
+            "header",
+            classes=["toga"],
+            children=[
+                create_element(
+                    "a",
+                    classes=["brand"],
+                    children=[
+                        create_element(
+                            "img",
+                            src="static/logo-32.png",
+                            alt="Logo",
+                            loading="lazy",
+                        )
+                    ],
+                ),
+                self.native_menu_container,
+                self.native_help_menu_container,
+            ],
+        )
+
+    def set_app(self, app):
+        super().set_app(app)
+        app.native.insertBefore(self.native_titlebar, self.native)
+
+    def _create_submenu(self, group, items):
+        submenu = create_element(
+            "sl-dropdown",
+            children=[
+                create_element(
+                    "span",
+                    id=f"menu-{id(group)}",
+                    classes=["menu"],
+                    slot="trigger",
+                    content=group.text,
+                ),
+                create_element(
+                    "sl-menu",
+                    children=items,
+                ),
+            ],
+        )
+        return submenu
+
+    def create_menus(self):
+        self._menu_groups = {}
+        submenu = None
+
+        for cmd in self.interface.app.commands:
+            if isinstance(cmd, Separator):
+                # TODO - add a section break
+                pass
+            else:
+                # TODO - this doesn't handle submenus properly;
+                # all menu items will appear in their root group.
+                submenu = self._menu_groups.setdefault(cmd.group, [])
+
+                menu_item = create_element(
+                    "sl-menu-item",
+                    content=cmd.text,
+                    disabled=not cmd.enabled,
+                )
+                menu_item.onclick = cmd._impl.dom_click
+
+                submenu.append(menu_item)
+
+        menu_container = create_element("nav", classes=["menubar"])
+        help_menu_container = create_element("nav", classes=["menubar"])
+
+        # If there isn't an explicit app menu group, add an inert placeholder
+        if toga.Group.APP not in self._menu_groups:
+            menu_container.appendChild(
+                create_element(
+                    "span",
+                    classes=["app"],
+                    content=self.interface.app.formal_name,
+                )
+            )
+
+        for group, items in self._menu_groups.items():
+            submenu = self._create_submenu(group, items)
+            if group != toga.Group.HELP:
+                menu_container.appendChild(submenu)
+            else:
+                help_menu_container.appendChild(submenu)
+
+        self.native_menu_container.replaceWith(menu_container)
+        self.native_menu_container = menu_container
+
+        self.native_help_menu_container.replaceWith(help_menu_container)
+        self.native_help_menu_container = help_menu_container
+
+    def create_toolbar(self):
+        self.interface.factory.not_implemented("Window.create_toolbar()")
